Menu service - menu-category api all 85% tested

+++ menu-service/src/main/java/com/restaurant/menu_service/service/menu/impl/MenuCategoryServiceImpl.java
@@ -51,8 +51,12 @@ public class MenuCategoryServiceImpl implements MenuCategoryService {
     }
 
     @Override
-    public void delete(Long id) {
+    public boolean delete(Long id) {
+        if (!repository.existsById(id)) {
+            return false;
+        }
         repository.deleteById(id);
+        return true;
     }
 }
